% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_oracle_cxn.R
\name{make_oracle_cxn}
\alias{make_oracle_cxn}
\title{make_oracle_cxn}
\usage{
make_oracle_cxn(fn.oracle.username = "_none_",
  fn.oracle.password = "_none_", fn.oracle.dsn = "_none_")
}
\arguments{
\item{fn.oracle.username}{default is \code{'_none_'} This is your username for
accessing oracle objects. If you have a value for this stored in your
environment (e.g. from an rprofile file), this can be left and that value will
be used.  If a value for this is provided, it will take priority over your
existing value.}

\item{fn.oracle.password}{default is \code{'_none_'} This is your password for
accessing oracle objects. If you have a value for this stored in your
environment (e.g. from an rprofile file), this can be left and that value will
be used.  If a value for this is provided, it will take priority over your
existing value.}

\item{fn.oracle.dsn}{default is \code{'_none_'} This is your dsn/ODBC
identifier for accessing oracle objects. If you have a value for this stored in your
environment (e.g. from an rprofile file), this can be left and that value will
be used.  If a value for this is provided, it will take priority over your
existing value.}
}
\description{
This function contains the logic for creating a connection to
Oracle.
}
\note{
This is stolen directly from the get_data() function of bio.datawrangling
(also created by me), but references to roracle were removed
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
