% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHailInRpt.R
\name{makeHailInRpt}
\alias{makeHailInRpt}
\title{makeHailInRpt}
\usage{
makeHailInRpt(thePath = file.path("C:", "DFO-MPO", "PORTSAMPLING"),
  fn.oracle.username = "_none_", fn.oracle.password = "_none_",
  fn.oracle.dsn = "_none_")
}
\arguments{
\item{thePath}{default is \code{C:\\DFO-MPO\\PORTSAMPLING}.  This is where you
would like your reports saved}

\item{fn.oracle.username}{default is \code{'_none_'} This is your username for
accessing oracle objects. If you have a value for this stored in your
environment (e.g. from an rprofile file), this can be left and that value will
be used.  If a value for this is provided, it will take priority over your
existing value.}

\item{fn.oracle.password}{default is \code{'_none_'} This is your password for
accessing oracle objects. If you have a value for this stored in your
environment (e.g. from an rprofile file), this can be left and that value will
be used.  If a value for this is provided, it will take priority over your
existing value.}

\item{fn.oracle.dsn}{default is \code{'_none_'} This is your dsn/ODBC
identifier for accessing oracle objects. If you have a value for this stored in your
environment (e.g. from an rprofile file), this can be left and that value will
be used.  If a value for this is provided, it will take priority over your
existing value.}
}
\description{
This function generates an excel file where the first sheet
lists all the groundfish hailins for the current and next day, and the
subsequent sheets include breakdowns of the catch.
}
\seealso{
Other portsampling: \code{\link{makePortSamplingBat}},
  \code{\link{update.portsampling}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
