% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePortSamplingBat.R
\name{makePortSamplingBat}
\alias{makePortSamplingBat}
\title{makePortSamplingBat}
\usage{
makePortSamplingBat()
}
\description{
This function facilitates the generation of a batch file to your
desktop that can be double-clicked to run reports easily
}
\note{
This is stolen directly from the get_data() function of bio.datawrangling
(also created by me), but references to roracle were removed
}
\seealso{
Other portsampling: \code{\link{makeHailInRpt}},
  \code{\link{update.portsampling}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
