% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePushButton.R
\name{makePushButton}
\alias{makePushButton}
\title{makePushButton}
\usage{
makePushButton(arch = NULL)
}
\arguments{
\item{arch}{default is \code{NULL}.  This is the r architecture you want to use - either 32 or 64 are 
valid entries.  If left as NULL, it will use your default architecture}
}
\description{
This function facilitates the generation of a batch file to your
desktop that can be double-clicked to run reports easily
}
\seealso{
Other Mar.portsampling: 
\code{\link{makeHailInRpt}()},
\code{\link{updatePortsampling}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{Mar.portsampling}
