% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.portsampling.R
\name{update.portsampling}
\alias{update.portsampling}
\title{update.portsampling}
\usage{
\method{update}{portsampling}()
}
\description{
This is just a quick and dirty script for facilitating updates
}
\seealso{
Other portsampling: \code{\link{makeHailInRpt}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
