% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePortsampling.R
\name{updatePortsampling}
\alias{updatePortsampling}
\title{updatePortsampling}
\usage{
updatePortsampling()
}
\description{
This is just a quick and dirty script for facilitating updates
}
\seealso{
Other portsampling: \code{\link{makeHailInRpt}},
  \code{\link{makePushButton}}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
