% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePushButton.R
\name{makePushButton}
\alias{makePushButton}
\title{makePushButton}
\usage{
makePushButton()
}
\description{
This function facilitates the generation of a batch file to your
desktop that can be double-clicked to run reports easily
}
\seealso{
Other Mar.portsampling: 
\code{\link{makeHailInRpt}()},
\code{\link{updatePortsampling}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{Mar.portsampling}
